Fix jvm cache manager

+++ composeApp/src/jvmMain/kotlin/org/sternbach/software/julesmobileapp/CacheManager.kt
@@ -3,8 +3,12 @@ package org.sternbach.software.julesmobileapp
 import java.io.File
 
 actual object CacheManager {
+    private fun getFile(key: String): File {
+        return File(System.getProperty("user.home"), ".jules_pref_$key")
+    }
+
     actual fun readPreference(key: String): String? {
-        val file = File(System.getProperty("user.home"), ".jules_$key")
+        val file = getFile(key)
         return if (file.exists()) {
             try {
                 file.readText()
@@ -17,21 +21,23 @@ actual object CacheManager {
     }
 
     actual fun writePreference(key: String, value: String?) {
-        val file = File(System.getProperty("user.home"), ".jules_$key")
-        if (value == null) {
-            file.delete()
-        } else {
-            try {
+        val file = getFile(key)
+        try {
+            if (value == null) {
+                if (file.exists()) {
+                    file.delete()
+                }
+            } else {
                 file.writeText(value)
-            } catch (e: Exception) {
-                e.printStackTrace()
             }
+        } catch (e: Exception) {
+            e.printStackTrace()
         }
     }
 
     actual fun readBooleanPreference(key: String, defaultValue: Boolean): Boolean {
         val str = readPreference(key)
-        return str?.toBoolean() ?: defaultValue
+        return str?.toBooleanStrictOrNull() ?: defaultValue
     }
 
     actual fun writeBooleanPreference(key: String, value: Boolean) {
